requirements.txt
certifi==2024.2.2
colorama==0.4.6
coverage==7.4.3
distlib==0.3.8
exceptiongroup==1.2.0
filelock==3.13.1
filetype==1.2.0
greenlet==3.0.3
iniconfig==2.0.0
Mako==1.3.2
MarkupSafe==2.1.5
mutagen==1.47.0
numpy==1.26.4
packaging==23.2
pandas==2.2.1
platformdirs==4.2.0
pluggy==1.4.0
psycopg2==2.9.9
pytest==8.0.1
pytest-cov==4.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.27
tinytag==1.10.1
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2024.1
virtualenv==20.25.0
